Api_cancer/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow==10.4.0
torch==2.4.1
torchvision==0.19.1
python-multipart==0.0.9